% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{capture_internal_function_def}
\alias{capture_internal_function_def}
\title{Captures whether an R language statement is the assignment of a function
call.}
\usage{
capture_internal_function_def(line, function_name, cache_this)
}
\arguments{
\item{line}{An R statement}

\item{function_name}{The name of the function that this line is in.}

\item{cache_this}{An environment to cache the results so this does not need
to be a two-pass parser}
}
\description{
Captures whether an R language statement is the assignment of a function
call.
}
