% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_dev_tools.R
\name{diagnose_werner_failures}
\alias{diagnose_werner_failures}
\title{Diagnoses werner's failures to detect foreign functions.}
\usage{
diagnose_werner_failures(package_name)
}
\arguments{
\item{package_name}{Name of a package to explore}
}
\description{
This function paradoxically subsets werner's returns to only include
function calls it specifically could not resolve; the intent here is to aid
werner package development by identifying edge cases that are not yet
supported.
}
