% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{explore_package}
\alias{explore_package}
\title{Maps relationships between functions in a package.}
\usage{
explore_package(package_name)
}
\arguments{
\item{package_name}{a character string containing the name of an installed package}
}
\value{
A list of each function in the package, each either NULL or
a vector of character strings referring to which other functions the function
calls.
}
\description{
Maps relationships between functions in a package.
}
