% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_calls.R
\name{get_calls_from_function}
\alias{get_calls_from_function}
\title{Gets all calls from a function}
\usage{
get_calls_from_function(function_name, all_functions, package_name, cache_this)
}
\arguments{
\item{function_name}{Character vector name of a function, namespaced}

\item{all_functions}{Vector of character strings listing all functions in
the package's namespace}

\item{package_name}{Name of the package we're searching}

\item{cache_this}{Environent to hold the progress bar and memoised progress}
}
\description{
The core workhorse of werner, this function gets all calls made from a given
function in a package by breaking the function call into a series of
discrete lines and handing each line off to `get_calls_from_line`.
}
