% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_functions_from_package}
\alias{get_functions_from_package}
\title{Extracts all the functions from a package}
\usage{
get_functions_from_package(package_name, cache_this)
}
\arguments{
\item{package_name}{Character vector name of a package}

\item{cache_this}{an environment -- used to memoise the function.}
}
\value{
Vector of character strings containing all the functions in the
package, namedspaced appropriately.
}
\description{
Loads a package, gets all the functions in the package, cache the result.
}
