% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{name_need_quote}
\alias{name_need_quote}
\title{Determines whether an R object name needs to be backticked for to be used.}
\usage{
name_need_quote(text)
}
\arguments{
\item{text}{Function name to check}
}
\value{
A character vector with the fixed version of the text
}
\description{
This is a simple wrapper for two cases where function names need to be
backticked -- first, if the name is a reserved keyword, second if the name
has certain symbols in its name.
}
