% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_adjacency_matrix}
\alias{plot_adjacency_matrix}
\title{Plotting a package adjacency matrix.}
\usage{
plot_adjacency_matrix(adj_matrix, col_called = "skyblue2",
  col_caller = "darkgoldenrod1", col_orphan = "lightgrey",
  space_multiplier = 1, ...)
}
\arguments{
\item{adj_matrix}{An adjacency matrix generated by `adjacency_matrix()`}

\item{col_called}{Graphical parameter for the color of functions that are
called by other functions, defaults to "skyblue2"}

\item{col_caller}{Graphical parameter for the color of functions that only
call other functions, defaults to "darkgoldenrod1".}

\item{col_orphan}{Graphical parameter for the color of functions that neither
call nor are called, defaults to "lightgrey"}

\item{space_multiplier}{Numeric that affects how much space between nodes in
the plot.}

\item{...}{Other graphical arguments, see `?qgraph` for more information.}
}
\description{
This function provides helpful defaults for plotting a package adjacency
matrix. The default settings attempt to maximize the space between dense
parts of the graph. Private, non-exported functions are prefixed by a
":" in the graph labels.
}
