% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/werner.R
\docType{package}
\name{werner}
\alias{werner}
\alias{werner-package}
\title{werner package}
\description{
werner tells you which functions in your package call which other functions so you can better understand your package's layout.
}
