% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{which_are_functions}
\alias{which_are_functions}
\title{Confirm which objects from a package namespace are functions}
\usage{
which_are_functions(package_name, candidate_functions, public)
}
\arguments{
\item{package_name}{The name of the package we're checking; needed to prefix
the namespace of the function call.}

\item{candidate_functions}{Vector of character vectors for function names
we're checking.}

\item{public}{Determines how many colons to put in the}
}
\value{
A vector of booleans for whether or not each item is a function.
}
\description{
This is a simple wrapper for checking if everything in a package namespace
is a function or something else.
}
