% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{adjacency_matrix}
\alias{adjacency_matrix}
\title{Generates an adjacency matrix between functions in a package}
\usage{
adjacency_matrix(package_name, coerce_to_matrix = FALSE)
}
\arguments{
\item{package_name}{a character string containing the name of an installed package}

\item{coerce_to_matrix}{If TRUE, return an R matrix instead of a sparse Matrix.}
}
\value{
A square matrix where each row and column are functions in the package.
The matrix is an adjacency function: a 1 in a cell indicates that the function named
in the row calls the function named in the column.
}
\description{
Generates an adjacency matrix between functions in a package
}
